feat: move ResourceAccordion into common
