+++ code/workspaces/web-app/src/components/common/ResourceAccordion.js
