+++ code/workspaces/web-app/src/containers/adminListUsers/ResourceStackCards.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Typography from '@material-ui/core/Typography';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import { ResourceAccordion, ResourceAccordionSummary, ResourceAccordionDetails } from '../adminResources/ResourceAccordion';
+import { ResourceAccordion, ResourceAccordionSummary, ResourceAccordionDetails } from '../../components/common/ResourceAccordion';
 import StackCards from '../../components/stacks/StackCards';
 
 function ResourceStackCards({ resources, typeName, typeNamePlural }) {
