+++ code/workspaces/web-app/src/containers/adminResources/ProjectNotebooks.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Typography from '@material-ui/core/Typography';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import { ResourceAccordion, ResourceAccordionSummary, ResourceAccordionDetails } from './ResourceAccordion';
+import { ResourceAccordion, ResourceAccordionSummary, ResourceAccordionDetails } from '../../components/common/ResourceAccordion';
 import ProjectNotebooksContainer from './ProjectNotebooksContainer';
 
 export default function ProjectNotebooks(props) {
