+++ code/workspaces/web-app/src/containers/adminResources/ProjectStorage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Typography from '@material-ui/core/Typography';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import { ResourceAccordion, ResourceAccordionSummary, ResourceAccordionDetails } from './ResourceAccordion';
+import { ResourceAccordion, ResourceAccordionSummary, ResourceAccordionDetails } from '../../components/common/ResourceAccordion';
 import { ProjectDataStorageContainer } from '../dataStorage/DataStorageContainer';
 
 export default function ProjectStorage(props) {
